Remove support for legacy connection handling
Simplifies the gem, ideally making it easier to get it working with Rails 7.1.
