+++ .github/workflows/ci.yml
@@ -15,7 +15,7 @@ jobs:
 
   tests:
     runs-on: ubuntu-latest
-    name: Test ${{ matrix.gemfile }} with Ruby ${{ matrix.ruby-version }} (LCH ${{ matrix.legacy_connection_handling }})
+    name: Test ${{ matrix.gemfile }} with Ruby ${{ matrix.ruby-version }}
     services:
       mysql:
         image: mysql:8.0
@@ -42,15 +42,8 @@ jobs:
           - rails7.0_trilogy
           - rails7.1_mysql2
           - rails7.1_trilogy
-        legacy_connection_handling:
-          - "false"
-          - "true"
-        exclude:
-          - {gemfile: rails7.1_mysql2,  legacy_connection_handling: "true"}
-          - {gemfile: rails7.1_trilogy, legacy_connection_handling: "true"}
     env:
       BUNDLE_GEMFILE: gemfiles/${{ matrix.gemfile }}.gemfile
-      LEGACY_CONNECTION_HANDLING: "${{ matrix.legacy_connection_handling }}"
     steps:
       - uses: actions/checkout@v4
       - name: Install Ruby, Bundler and gems
