+++ test/test_thread_safety.rb
@@ -6,11 +6,7 @@ class ThreadSafetyTest < Minitest::Test
   include ARHPTestSetup
 
   def setup
-    if LEGACY_CONNECTION_HANDLING
-      Phenix.rise!
-    else
-      Phenix.rise! config_path: "test/three_tier_database.yml"
-    end
+    Phenix.rise! config_path: "test/three_tier_database.yml"
 
     arhp_create_models
